Issue #20 - Drop support for interval
In some cases, if we call for Performance data for few signatures and an
interval we will get a 503.

See bug https://bugzilla.mozilla.org/show_bug.cgi?id=1535432 for all
details.

In this change we drop the support to appending `interval` to any URLs.
This will prevent hitting the bug.

The burden of showing less data points will be shifted to the frontend.